Restore old job so existing jobs in queue don't fail
