+++ app/Jobs/EsIndexDocument.php
@@ -5,17 +5,21 @@
 
 namespace App\Jobs;
 
-use Exception;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\SerializesModels;
 
+/**
+ * TODO: This job has been replaced with EsDocument.
+ * It's left around temporarily so existing job can finish processing.
+ */
 class EsIndexDocument implements ShouldQueue
 {
-    use Dispatchable, InteractsWithQueue, Queueable;
+    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
-    private array $modelMeta;
+    protected $model;
 
     /**
      * Create a new job instance.
@@ -24,10 +28,7 @@ class EsIndexDocument implements ShouldQueue
      */
     public function __construct($model)
     {
-        $this->modelMeta = [
-            'class' => get_class($model),
-            'id' => $model->getKey(),
-        ];
+        $this->model = $model;
     }
 
     /**
@@ -35,28 +36,8 @@ class EsIndexDocument implements ShouldQueue
      *
      * @return void
      */
-    public function handle(): void
+    public function handle()
     {
-        $id = $this->modelMeta['id'];
-
-        if (!is_scalar($id)) {
-            log_error(new Exception("can't index models with unsupported primary key: ".json_encode($id)));
-
-            return;
-        }
-
-        $class = $this->modelMeta['class'];
-        $model = $class::find($id);
-
-        if ($model !== null) {
-            $model->esIndexDocument();
-
-            return;
-        }
-
-        $model = new $class();
-        $keyName = $model->getKeyName();
-        $model->setAttribute($keyName, $id);
-        $model->esDeleteDocument();
+        $this->model->esIndexDocument();
     }
 }
